REF: move _egg_info tests into separate module.
